ticker: document it's unsafe to access backoff policy while ticker running
The other way around would be to use locks to make backoff policies
thread-safe. This is not something hard to do but this would need to
be implemented in each policy. It doesn't seem unreasonable to require
the user to not modify the backoff policy while we are using it.

Fix #47